+++ webpack.prod.js
@@ -1,19 +1,25 @@
 const path = require('path');
 const merge = require('webpack-merge');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
-const paths = require('./paths');
+const TerserPlugin = require('terser-webpack-plugin');
 const common = require('./webpack.common');
-
-const BUILD = path.join(paths.BASE, 'dist');
 
 const CleanWebpackPluginConfig = new CleanWebpackPlugin(['dist']);
 
 module.exports = merge(common, {
   output: {
-    path: BUILD,
-    filename: 'bundle.js'
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'bundle.min.js'
   },
   plugins: [
     CleanWebpackPluginConfig
-  ]
+  ],
+  optimization: {
+    minimize: true,
+    minimizer: [
+      new TerserPlugin({
+        sourceMap: true
+      })
+    ],
+  }
 });